// src/molecules/Dropdown/Dropdown.types.ts
// Generated with util/create-component.js
export type dropdownGroupPositionType = 'left' | 'center' | 'right' | undefined;

export interface DropdownProps {
	category: string;
	titles: string[];
	mobile?: boolean;
	open?: boolean;
	hoverColor?: string;
	inGroupPosition?: dropdownGroupPositionType;
}